requirements.txt
fastapi
uvicorn[standard]
google-generativeai
playwright
beautifulsoup4
httpx
aiofiles
python-dotenv
anyio
nest-asyncio
lxml